skin_burn_app/requirements.txt
flask
werkzeug
pillow
numpy
tensorflow
keras
lime
scikit-image
matplotlib
opencv-python
reportlab
